+++ src/Interfaces/ViewOrActionResultInterface.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * This file is a part of the bluemvc-core package.
+ *
+ * Read more at https://bluemvc.com/
+ */
+declare(strict_types=1);
+
+namespace BlueMvc\Core\Interfaces;
+
+/**
+ * Interface for View or ActionResult class.
+ *
+ * @since 2.1.0
+ */
+interface ViewOrActionResultInterface
+{
+}