+++ tests/Helpers/TestControllers/BasicTestController.php
@@ -4,10 +4,16 @@ declare(strict_types=1);
 
 namespace BlueMvc\Core\Tests\Helpers\TestControllers;
 
+use BlueMvc\Core\ActionResults\NoContentResult;
+use BlueMvc\Core\ActionResults\NotModifiedResult;
 use BlueMvc\Core\Controller;
 use BlueMvc\Core\Http\StatusCode;
+use BlueMvc\Core\Interfaces\ActionResults\ActionResultInterface;
+use BlueMvc\Core\Interfaces\ViewInterface;
+use BlueMvc\Core\Interfaces\ViewOrActionResultInterface;
 use BlueMvc\Core\Tests\Helpers\TestClasses\SimpleTestClass;
 use BlueMvc\Core\Tests\Helpers\TestClasses\StringableTestClass;
+use BlueMvc\Core\View;
 
 /**
  * Basic test controller class.
@@ -19,7 +25,7 @@ class BasicTestController extends Controller
      *
      * @return string The result.
      */
-    public function indexAction()
+    public function indexAction(): string
     {
         return 'Hello World!';
     }
@@ -29,7 +35,7 @@ class BasicTestController extends Controller
      *
      * @return string The result.
      */
-    public function serverErrorAction()
+    public function serverErrorAction(): string
     {
         $this->getResponse()->setStatusCode(new StatusCode(StatusCode::INTERNAL_SERVER_ERROR));
 
@@ -41,7 +47,7 @@ class BasicTestController extends Controller
      *
      * @return string The result.
      */
-    public function _123numericAction()
+    public function _123numericAction(): string
     {
         return 'Numeric action result';
     }
@@ -51,7 +57,7 @@ class BasicTestController extends Controller
      *
      * @return int The result.
      */
-    public function intAction()
+    public function intAction(): int
     {
         return 42;
     }
@@ -61,7 +67,7 @@ class BasicTestController extends Controller
      *
      * @return bool The result.
      */
-    public function falseAction()
+    public function falseAction(): bool
     {
         return false;
     }
@@ -71,7 +77,7 @@ class BasicTestController extends Controller
      *
      * @return bool The result.
      */
-    public function trueAction()
+    public function trueAction(): bool
     {
         return true;
     }
@@ -93,7 +99,7 @@ class BasicTestController extends Controller
      *
      * @return SimpleTestClass The result.
      */
-    public function objectAction()
+    public function objectAction(): SimpleTestClass
     {
         return new SimpleTestClass('Foo');
     }
@@ -103,8 +109,42 @@ class BasicTestController extends Controller
      *
      * @return StringableTestClass The result.
      */
-    public function stringableAction()
+    public function stringableAction(): StringableTestClass
     {
         return new StringableTestClass('Bar');
+    }
+
+    /**
+     * Action returning an action result.
+     *
+     * @return ActionResultInterface The result.
+     */
+    public function actionResultAction(): ActionResultInterface
+    {
+        return new NotModifiedResult();
+    }
+
+    /**
+     * Action returning a view.
+     *
+     * @return ViewInterface The result.
+     */
+    public function viewAction(): ViewInterface
+    {
+        return new View();
+    }
+
+    /**
+     * Action returning a view or action result.
+     *
+     * @return ViewOrActionResultInterface The result.
+     */
+    public function viewOrActionResultAction(): ViewOrActionResultInterface
+    {
+        if ($this->getRequest()->getQueryParameter('showView') === '1') {
+            return new View();
+        }
+
+        return new NoContentResult();
     }
 }